backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
flask-cors==5.0.1
google-generativeai==0.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.1.0
pypdf==5.3.1
python-dotenv==1.0.1
typing_extensions==4.12.2
Werkzeug==3.1.3
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1